# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gavhah Community Kindness Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /discussions:
    get:
      operationId: listDiscussions
      tags: [discussions]
      summary: List forum discussions
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: trending
          schema:
            type: boolean
        - in: query
          name: page
          schema:
            type: integer
      responses:
        "200":
          description: List of discussions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discussion"
    post:
      operationId: createDiscussion
      tags: [discussions]
      summary: Create a new discussion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscussionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discussion"
  /discussions/{id}:
    get:
      operationId: getDiscussion
      tags: [discussions]
      summary: Get discussion by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Discussion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discussion"
    patch:
      operationId: updateDiscussion
      tags: [discussions]
      summary: Update a discussion
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscussionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discussion"
    delete:
      operationId: deleteDiscussion
      tags: [discussions]
      summary: Delete a discussion
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /discussions/{id}/like:
    post:
      operationId: likeDiscussion
      tags: [discussions]
      summary: Like a discussion
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
  /discussions/{id}/lock:
    post:
      operationId: lockDiscussion
      tags: [discussions]
      summary: Lock a discussion (moderator/admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discussion"
  /discussions/{id}/comments:
    get:
      operationId: listDiscussionComments
      tags: [discussions]
      summary: List comments for a discussion
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createDiscussionComment
      tags: [discussions]
      summary: Comment on a discussion
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List community groups
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: privacy
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update a group
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /groups/{id}/join:
    post:
      operationId: joinGroup
      tags: [groups]
      summary: Request to join a group
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Join request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMember"
  /groups/{id}/members:
    get:
      operationId: listGroupMembers
      tags: [groups]
      summary: List group members
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupMember"
  /groups/{id}/posts:
    get:
      operationId: listGroupPosts
      tags: [groups]
      summary: List posts in a group
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupPost"
    post:
      operationId: createGroupPost
      tags: [groups]
      summary: Create a post in a group
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupPost"

  # ---- MINYANS ----
...
  /charity/{id}:
    get:
      operationId: getCharity
      tags: [charity]
      summary: Get charity by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Charity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charity"
    patch:
      operationId: updateCharity
      tags: [charity]
      summary: Update a charity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charity"
  /charity/{id}/donate:
    post:
      operationId: donateToCharity
      tags: [charity]
      summary: Record a donation intent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonationInput"
      responses:
        "201":
          description: Donation recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"

  # ---- STATS ----
...
components:
  schemas:
...
    Discussion:
      type: object
      required: [id, title, content, category, authorId, authorName, likes, views, commentCount, isLocked, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [medical, shidduchim, livelihood, education, charity, community, general]
        authorId:
          type: integer
        authorName:
          type: string
        likes:
          type: integer
        views:
          type: integer
        commentCount:
          type: integer
        isLocked:
          type: boolean
        isPinned:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    DiscussionInput:
      type: object
      required: [title, content, category]
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
    DiscussionUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        isPinned:
          type: boolean
    Comment:
      type: object
      required: [id, content, authorId, authorName, likes, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        parentId:
          type: ["integer", "null"]
        likes:
          type: integer
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        parentId:
          type: ["integer", "null"]
    LikeResult:
      type: object
      required: [likes]
      properties:
        likes:
          type: integer

    # ---- NEWS ----
...
    Group:
      type: object
      required: [id, name, description, privacy, ownerId, ownerName, memberCount, postCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        privacy:
          type: string
          enum: [public, private, password_protected]
        ownerId:
          type: integer
        ownerName:
          type: string
        imageUrl:
          type: ["string", "null"]
        memberCount:
          type: integer
        postCount:
          type: integer
        createdAt:
          type: string
    GroupInput:
      type: object
      required: [name, description, privacy]
      properties:
        name:
          type: string
        description:
          type: string
        privacy:
          type: string
        password:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        privacy:
          type: string
        imageUrl:
          type: ["string", "null"]
    GroupMember:
      type: object
      required: [id, userId, userName, groupId, role, status, joinedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        groupId:
          type: integer
        role:
          type: string
          enum: [owner, moderator, member]
        status:
          type: string
          enum: [approved, pending]
        joinedAt:
          type: string
    GroupPost:
      type: object
      required: [id, content, groupId, authorId, authorName, likes, views, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        groupId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        likes:
          type: integer
        views:
          type: integer
        createdAt:
          type: string
    GroupPostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ---- MINYAN ----
...
    Charity:
      type: object
      required: [id, name, description, goalAmount, raisedAmount, isTodaysFeatured, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        successStories:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        goalAmount:
          type: integer
        raisedAmount:
          type: integer
        isTodaysFeatured:
          type: boolean
        createdAt:
          type: string
    CharityUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        successStories:
          type: string
        imageUrl:
          type: string
        goalAmount:
          type: integer
        isTodaysFeatured:
          type: boolean
    DonationInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
        donorName:
          type: ["string", "null"]
    Donation:
      type: object
      required: [id, charityId, amount, createdAt]
      properties:
        id:
          type: integer
        charityId:
          type: integer
        amount:
          type: integer
        donorName:
          type: ["string", "null"]
        createdAt:
          type: string

    # ---- STATS ----
...