# lib/api-zod/src/generated/types/groupInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gavhah Community Kindness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GroupInput {
  name: string;
  description: string;
  privacy: string;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}