# lib/api-zod/src/generated/types/groupPostInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gavhah Community Kindness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GroupPostInput {
  content: string;
}